Stitch a new method to set ResourceBundle.parent correctly.
